requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
langchain==0.1.6
langchain-core==0.1.27
google-generativeai==0.7.1
pydantic==2.6.3
